# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translations API",
        "description": "API Documentation for Translations",
        "contact": {
            "email": "aqibullah3312@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/translations": {
            "get": {
                "tags": [
                    "Translation"
                ],
                "summary": "Get Translations",
                "description": "Get translations",
                "operationId": "434a55804ca50c9845bd4781c222a539",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter translations by search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter translations with key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter translations by locale",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter translations by tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "locale": {
                                            "type": "string",
                                            "default": "en"
                                        },
                                        "group": {
                                            "type": "string",
                                            "example": "messages"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "hello_world"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "Hello World"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "mobile",
                                                "desktop"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translations not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Add Translation",
                "description": "Add a new translation",
                "operationId": "fe38d8cf02ff5c92d2cd8d5dfb095e5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "group",
                                    "key",
                                    "value",
                                    "tags"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "default": "en"
                                    },
                                    "group": {
                                        "type": "string",
                                        "example": "messages"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "hello_world"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Hello World"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "mobile",
                                            "desktop"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "locale": {
                                            "type": "string",
                                            "default": "en"
                                        },
                                        "group": {
                                            "type": "string",
                                            "example": "messages"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "hello_world"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "Hello World"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "mobile",
                                                "desktop"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translation"
                ],
                "summary": "Export Translations",
                "description": "Add a new translation",
                "operationId": "52e76cf9a0dbfcb42b1d151b49f2b09b",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "locale": {
                                            "type": "string",
                                            "default": "en"
                                        },
                                        "group": {
                                            "type": "string",
                                            "example": "messages"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "hello_world"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "Hello World"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "mobile",
                                                "desktop"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No translations found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}